creates adminOptions.js container and implements user and product administration
