+++ src/components/layouts/adminOptions/articleAdministration/productTable.js
@@ -19,7 +19,7 @@ import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
 import DeleteIcon from '@material-ui/icons/Delete';
 import FilterListIcon from '@material-ui/icons/FilterList';
-import {getProducts} from "../../connectionHandler/connectionHandler";
+import {getProducts} from "../../../connectionHandler/connectionHandler";
 
 
 const ProductTable = ({setErrorMessage, products, setProducts}) => {
